requirements.txt
# airflow users create \
#     --username admin \
#     --firstname Avazbek \
#     --lastname Isroilov \
#     --role Admin \
#     --email isroilovavazbek2002@gmail.com

# HTTP requests library
requests>=2.31.0

# BeautifulSoup for HTML parsing
beautifulsoup4>=4.12.0

# Regular expressions (part of Python standard library)
# re is built-in, no need to include

# CSV (part of Python standard library)
# csv is built-in, no need to include

# DateTime (part of Python standard library)
# datetime is built-in, no need to include

# Apache Airflow and its dependencies
apache-airflow>=2.7.0
apache-airflow-providers-http>=4.6.0
apache-airflow-providers-common-sql>=1.8.0
apache-airflow-providers-postgres>=5.7.0
apache-airflow-providers-ssh>=3.10.0
apache-airflow-providers-ftp>=3.6.0
apache-airflow-providers-imap>=3.4.0
apache-airflow-providers-sqlite>=3.5.0
markupsafe>=2.1.0
sqlalchemy>=1.4.0
pendulum>=2.1.0
python-dateutil>=2.8.0
croniter>=1.3.0
flask>=2.2.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.0
werkzeug>=2.2.0